requirements.txt
# Python 依赖
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.5
alembic==1.13.0
httpx==0.27.0
openai==1.50.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
redis==5.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
beautifulsoup4==4.12.3
